requirements.txt
streamlit
pandas
joblib
sklearn
Pillow